model-net-mpi-replay: fix is_synthetic locality bug
Before this, if a simulation was executed with multiple jobs, and
one of those background jobs was synthetic, and there existed a PE
with none of the synthetic ranks, then the background synthetic ranks
would never stop generating data and the simulation would not end.

The cause of this was the `is_synthetic` global static variable that
is set to 1 whenever a synthetic rank is initialized. This appears
to be a way to prevent needlessly calling notify_neighbors() when
a primary workload finishes. There's no point to calling that if
the other ranks are not synthetic. But since it's possible for a
PE to exist without any synthetic ranks, `is_synthetic` would be
equal to zero for any LP on that PE. The result is that when the
primary workload is completing, there is a conditional that checks
to see if `is_synthetic` is 0, if it is, then there is no
notification of a completed workload. Thus it will appear like
there are still ranks working on the primary workload even when
they've finished long ago.